@@ -11,28 +11,29 @@
     "require": {
         "php": "^8.2",
         "ext-bcmath": "*",
-        "inertiajs/inertia-laravel": "^2.0",
-        "laravel/fortify": "^1.30",
-        "laravel/framework": "^12.0",
-        "laravel/tinker": "^2.10.1",
-        "laravel/wayfinder": "^0.1.9",
-        "pusher/pusher-php-server": "^7.2"
+        "inertiajs/inertia-laravel": "^2.0.14",
+        "laravel/fortify": "^1.32.1",
+        "laravel/framework": "^12.42",
+        "laravel/tinker": "^2.10.2",
+        "laravel/wayfinder": "^0.1.12",
+        "pusher/pusher-php-server": "^7.2.7"
     },
     "require-dev": {
         "roave/security-advisories": "dev-latest"
         ,
-        "barryvdh/laravel-debugbar": "^3.16",
-        "driftingly/rector-laravel": "^2.1",
-        "fakerphp/faker": "^1.23",
-        "larastan/larastan": "^3.0",
-        "laravel/pail": "^1.2.2",
-        "laravel/pint": "^1.24",
-        "laravel/sail": "^1.41",
-        "mockery/mockery": "^1.6",
-        "nunomaduro/collision": "^8.6",
-        "pestphp/pest": "^4.1",
+        "barryvdh/laravel-debugbar": "^3.16.1",
+        "driftingly/rector-laravel": "^2.1.7",
+        "fakerphp/faker": "^1.24.1",
+        "larastan/larastan": "^3.8",
+        "laravel/pail": "^1.2.4",
+        "laravel/pint": "^1.26",
+        "laravel/sail": "^1.51",
+        "mockery/mockery": "^1.6.12",
+        "nunomaduro/collision": "^8.8.3",
+        "pestphp/pest": "^4.1.6",
         "pestphp/pest-plugin-laravel": "^4.0",
-        "rector/rector": "^2.2"
+        "pestphp/pest-plugin-type-coverage": "^4.0",
+        "rector/rector": "^2.2.14"
     },
     "autoload": {
         "psr-4": {
@@ -64,16 +65,14 @@
             "Composer\\Config::disableProcessTimeout",
             "npx concurrently -c \"#93c5fd,#c4b5fd,#fb7185,#fdba74\" \"php artisan serve\" \"php artisan queue:listen --tries=1\" \"php artisan pail --timeout=0\" \"php artisan inertia:start-ssr\" --names=server,queue,logs,ssr --kill-others"
         ],
-        "test": [
-            "@php artisan config:clear --ansi",
-            "@php artisan test"
-        ],
         "post-autoload-dump": [
             "Illuminate\\Foundation\\ComposerScripts::postAutoloadDump",
             "@php artisan package:discover --ansi"
         ],
         "post-update-cmd": [
-            "@php artisan vendor:publish --tag=laravel-assets --ansi --force"
+            "@php artisan vendor:publish --tag=laravel-assets --ansi --force",
+            "@php artisan clear",
+            "@update:requirements"
         ],
         "post-root-package-install": [
             "@php -r \"file_exists('.env') || copy('.env.example', '.env');\""
@@ -85,6 +84,30 @@
         ],
         "pre-package-uninstall": [
             "Illuminate\\Foundation\\ComposerScripts::prePackageUninstall"
+        ],
+        "update:requirements": [
+            "composer bump",
+            "npx npm-check-updates -u"
+        ],
+        "lint": [
+            "rector",
+            "pint --parallel",
+            "npm run lint"
+        ],
+        "test:type-coverage": "./vendor/bin/pest --type-coverage --min=100 --memory-limit=2G",
+        "test:coverage": "XDEBUG_MODE=coverage pest --parallel --coverage --ci --min=100.0",
+        "test:lint": [
+            "pint --parallel --test",
+            "rector --dry-run",
+            "npm run lint"
+        ],
+        "test:unit": "./vendor/bin/pest",
+        "test:types": "phpstan  --memory-limit=2G",
+        "test": [
+            "@test:type-coverage",
+            "@test:unit",
+            "@test:lint",
+            "@test:types"
         ]
     },
     "extra": {
