@@ -10,10 +10,13 @@ use Illuminate\Broadcasting\InteractsWithSockets;
 use Illuminate\Contracts\Broadcasting\ShouldBroadcastNow;
 use Illuminate\Foundation\Events\Dispatchable;
 use Illuminate\Queue\SerializesModels;
+use Illuminate\Support\Collection;
 
 final class OrderbookUpdated implements ShouldBroadcastNow
 {
-    use Dispatchable, InteractsWithSockets, SerializesModels;
+    use Dispatchable;
+    use InteractsWithSockets;
+    use SerializesModels;
 
     public function __construct(
         public string $symbol,
@@ -50,10 +53,10 @@ final class OrderbookUpdated implements ShouldBroadcastNow
             ->limit(20)
             ->get()
             ->groupBy('price')
-            ->map(fn ($orders) => [
-                'price' => $orders->first()->price,
+            ->map(fn (Collection $orders): array => [
+                'price' => $orders->first()?->price,
                 'amount' => $orders->sum('amount'),
-                'total' => $orders->sum(fn ($order) => bcmul((string) $order->price, (string) $order->amount, 8)),
+                'total' => $orders->sum(fn (Order $order): string => bcmul($order->price, $order->amount, 8)),
                 'count' => $orders->count(),
             ])
             ->values();
@@ -67,10 +70,10 @@ final class OrderbookUpdated implements ShouldBroadcastNow
             ->limit(20)
             ->get()
             ->groupBy('price')
-            ->map(fn ($orders) => [
-                'price' => $orders->first()->price,
+            ->map(fn (Collection $orders): array => [
+                'price' => $orders->first()?->price,
                 'amount' => $orders->sum('amount'),
-                'total' => $orders->sum(fn ($order) => bcmul((string) $order->price, (string) $order->amount, 8)),
+                'total' => $orders->sum(fn (Order $order): string => bcmul($order->price, $order->amount, 8)),
                 'count' => $orders->count(),
             ])
             ->values();
