@@ -4,13 +4,20 @@ declare(strict_types=1);
 
 namespace App\Models;
 
+use Database\Factories\TradeFactory;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 
+/**
+ * @property numeric-string $price
+ * @property numeric-string $amount
+ * @property numeric-string $total
+ * @property numeric-string $commission
+ */
 final class Trade extends Model
 {
-    /** @use HasFactory<\Database\Factories\TradeFactory> */
+    /** @use HasFactory<TradeFactory> */
     use HasFactory;
 
     protected $fillable = [
