@@ -1,21 +1,23 @@
 <?php
 
+declare(strict_types=1);
+
 use App\Models\User;
 use Inertia\Testing\AssertableInertia as Assert;
 
-test('confirm password screen can be rendered', function () {
+test('confirm password screen can be rendered', function (): void {
     $user = User::factory()->create();
 
     $response = $this->actingAs($user)->get(route('password.confirm'));
 
     $response->assertStatus(200);
 
-    $response->assertInertia(fn (Assert $page) => $page
+    $response->assertInertia(fn (Assert $page): Assert => $page
         ->component('auth/ConfirmPassword')
     );
 });
 
-test('password confirmation requires authentication', function () {
+test('password confirmation requires authentication', function (): void {
     $response = $this->get(route('password.confirm'));
 
     $response->assertRedirect(route('login'));
