@@ -6,7 +6,7 @@ use App\Models\User;
 use Illuminate\Auth\Notifications\VerifyEmail;
 use Illuminate\Support\Facades\Notification;
 
-test('sends verification notification', function () {
+test('sends verification notification', function (): void {
     Notification::fake();
 
     $user = User::factory()->unverified()->create();
@@ -18,7 +18,7 @@ test('sends verification notification', function () {
     Notification::assertSentTo($user, VerifyEmail::class);
 });
 
-test('does not send verification notification if email is verified', function () {
+test('does not send verification notification if email is verified', function (): void {
     Notification::fake();
 
     $user = User::factory()->create();
