@@ -4,12 +4,12 @@ declare(strict_types=1);
 
 use App\Models\User;
 
-test('guests are redirected to the login page', function () {
+test('guests are redirected to the login page', function (): void {
     $response = $this->get(route('trading'));
     $response->assertRedirect(route('login'));
 });
 
-test('authenticated users can visit the trading page', function () {
+test('authenticated users can visit the trading page', function (): void {
     $user = User::factory()->create();
     $this->actingAs($user);
 
