@@ -1,6 +1,8 @@
 <?php
 
-test('returns a successful response', function () {
+declare(strict_types=1);
+
+test('returns a successful response', function (): void {
     $response = $this->get(route('home'));
 
     $response->assertStatus(200);
